requirements.txt
pyserial>=3.5
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0